+++ src/views/mod.rs
@@ -1,4 +1,5 @@
 pub mod hardware;
 pub mod info;
+pub mod status;
 pub mod version;
 pub mod waveform;
